Fix almost the everything
Things fixed:
	Blocks now have textures
	Reduced compiled file size by using texture altas.
	Pass 1 now renders correctly.
	Water now renders.
Things not fixed
	Leaves still don't render.

+++ source/display_list.h
@@ -6,12 +6,12 @@ typedef struct {
 	u16 index;
 	s16 *vertex;
 	u16 *color;
-	u8 *texcoord;
+	f32 *texcoord;
 } displayList;
 
 displayList *displist_create(u16 size);
 void displist_start();
 void displist_clear(displayList *list);
 void displist_bind(displayList *list);
-void displist_add(s16 x, s16 y, s16 z, u16 c, u8 u, u8 v);
+void displist_add(s16 x, s16 y, s16 z, u16 c, f32 u, f32 v);
 void displist_render(displayList *list);
